+++ block/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @platforma-open/milaboratories.mixcr-clonotyping
 
+## 1.8.8
+
+### Patch Changes
+
+- Updated dependencies [a370729]
+  - @platforma-open/milaboratories.mixcr-clonotyping.model@1.7.0
+  - @platforma-open/milaboratories.mixcr-clonotyping.ui@1.9.0
+
+## 1.8.7
+
+### Patch Changes
+
+- Updated dependencies [db3cc0a]
+  - @platforma-open/milaboratories.mixcr-clonotyping.workflow@1.12.0
+  - @platforma-open/milaboratories.mixcr-clonotyping.model@1.6.0
+  - @platforma-open/milaboratories.mixcr-clonotyping.ui@1.8.0
+
 ## 1.8.6
 
 ### Patch Changes