+++ model/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @platforma-open/milaboratories.mixcr-clonotyping.model
 
+## 1.7.0
+
+### Minor Changes
+
+- a370729: Custom library species corrected
+
+## 1.6.0
+
+### Minor Changes
+
+- db3cc0a: Custom reference library from library builder block or file
+
 ## 1.5.1
 
 ### Patch Changes