+++ ui/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @platforma-open/milaboratories.mixcr-clonotyping.ui
 
+## 1.9.0
+
+### Minor Changes
+
+- a370729: Custom library species corrected
+
+### Patch Changes
+
+- Updated dependencies [a370729]
+  - @platforma-open/milaboratories.mixcr-clonotyping.model@1.7.0
+
+## 1.8.0
+
+### Minor Changes
+
+- db3cc0a: Custom reference library from library builder block or file
+
+### Patch Changes
+
+- Updated dependencies [db3cc0a]
+  - @platforma-open/milaboratories.mixcr-clonotyping.model@1.6.0
+
 ## 1.7.2
 
 ### Patch Changes